Added PDF guide from the Embedded Programming with Black Magic workshop.
The Embedded Programming with Black Magic workshop is a 1 hour self
guided class giving a quick dive into ARM Cortex embedded programming
using open source tools, command line, and no IDE abstraction. The goal
of the class is to show how easy it is to use the simple tools and
understand the whole stack of embedded programming.